requirements.txt
Django==2.1.4
django-filter==2.0.0
django-filters==0.2.1
djangorestframework==3.9.0
pkg-resources==0.0.0
pytz==2018.7
